+++ crev-common/src/lib.rs
@@ -10,6 +10,9 @@ pub mod rand;
 pub mod serde;
 
 pub use crate::blake2b256::Blake2b256;
+
+use base64::engine::general_purpose::URL_SAFE_NO_PAD;
+use base64::Engine;
 use blake2::{digest::FixedOutput, Digest};
 use std::{
     collections::HashSet,
@@ -52,11 +55,11 @@ pub fn blake2b256sum_file(path: &Path) -> io::Result<[u8; 32]> {
 }
 
 pub fn base64_decode<T: ?Sized + AsRef<[u8]>>(input: &T) -> Result<Vec<u8>, base64::DecodeError> {
-    base64::decode_config(input, base64::URL_SAFE_NO_PAD)
+    URL_SAFE_NO_PAD.decode(input)
 }
 
 pub fn base64_encode<T: ?Sized + AsRef<[u8]>>(input: &T) -> String {
-    base64::encode_config(input, base64::URL_SAFE_NO_PAD)
+    URL_SAFE_NO_PAD.encode(input)
 }
 
 /// Takes a name and converts it to something safe for use in paths etc.