+++ flake.nix
@@ -2,8 +2,8 @@
   description = "Cryptographically verifiable Code REviews";
 
   inputs = {
-    nixpkgs.url = "github:NixOS/nixpkgs/nixos-24.05";
-    crane.url = "github:ipetkov/crane";
+    nixpkgs.url = "github:NixOS/nixpkgs/nixos-25.05";
+    crane.url = "github:ipetkov/crane/v0.18.1";
     crane.inputs.nixpkgs.follows = "nixpkgs";
     flake-utils.url = "github:numtide/flake-utils";
     fenix = {
@@ -24,7 +24,7 @@
         };
         lib = pkgs.lib;
 
-        fenix-channel = fenix.packages.${system}.latest;
+        fenix-channel = fenix.packages.${system}.stable;
 
         fenix-toolchain = (fenix-channel.withComponents [
           "rustc"
@@ -34,7 +34,9 @@
           "llvm-tools-preview"
         ]);
 
-        craneLib = (crane.mkLib pkgs).overrideToolchain fenix-toolchain;
+        craneLib = (crane.mkLib pkgs).overrideToolchain (p:
+          fenix.packages.${system}.fromManifestFile ./rust-toolchain.toml
+        );
 
         # filter source code at path `src` to include only the list of `modules`
         filterModules = modules: src:
@@ -228,7 +230,7 @@
 
                 pkgs.nixpkgs-fmt
                 pkgs.shellcheck
-                pkgs.rnix-lsp
+                pkgs.nil
                 pkgs.nodePackages.bash-language-server
               ]);
             RUST_SRC_PATH = "${fenix-channel.rust-src}/lib/rustlib/src/rust/library";