+++ cargo-crev/CHANGELOG.md
@@ -3,6 +3,7 @@
 <!-- next-url -->
 ## [Unreleased](https://github.com/crev-dev/cargo-crev/compare/v0.26.4...HEAD) - ReleaseDate
 
+- Added mac keychain support to safely store and retrieve passphrases
 
 ## [0.26.4](https://github.com/crev-dev/cargo-crev/compare/v0.26.0...v0.26.4) - 2025-04-12